backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.4
